enum.py renamed to _enum.py to avoid name clash with the built-in `enum` module while running shapely modules in debug mode https://github.com/shapely/shapely/issues/1693
