+++ shapely/_enum.py
