+++ shapely/_geometry.py
@@ -4,8 +4,8 @@ from enum import IntEnum
 import numpy as np
 
 from . import _geometry_helpers, geos_version, lib
+from ._enum import ParamEnum
 from .decorators import multithreading_enabled, requires_geos
-from .enum import ParamEnum
 
 __all__ = [
     "GeometryType",
